requirements.txt
fastapi
uvicorn
sqlmodel
pydantic-settings
python-jose
